Added updateKey and getJSON()
Everytime a match is updated drives and plays will get an updateKey, which allows a client to look for everything that changed since an updateKey from the past.
Also every class now contains a function that returns every needed data of the match as a JavaScript object. This is useful for future development :)